% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{aggregate_treelist}
\alias{aggregate_treelist}
\title{Get Species Composition}
\usage{
aggregate_treelist(treelist)
}
\arguments{
\item{treelist}{A treelist - a tidy dataframe with a row for each observation
of an individual tree (a single tree may have multiple rows if a plot was
visited more than once), and columns for plot_id, inv_date, species, and
dbh_cm. See "help(validate_treelist)" for more details.}
}
\value{
A tidy data frame with a row for each observation of a sampling
location (a unique combination of plot_id and inv_date), and columns for
the proportion of basal area occupied by each species in the dataset for
each observation. The proportion of plot basal area occupied by all species
('pBA_(all)') should always be equal to 1.
}
\description{
For calculating fuel loads, we want to know the species composition of the
local overstory. 'Local' means the overstory composition which is
representative of the actual composition of surface fuels on the plot.
Often, field data will include a treelist collected from the same plot
(at the same time) as the Brown's transects. We can use this treelist to
describe the species composition for each plot, and use the plot-specific
species composition numbers to estimate the fuel load. 'Plot' generally
means a sampling unit (e.g. a 0.04-ha sampling plot), but users could
use stand-level overstory data by feeding a stand identifier into the
plot_id field (as long as fuels data have the same set of stand
identifiers). We express species as the proportion of total basal area
occupied by each species.
}
