% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fuel_loads.R
\name{estimate_cwd_load}
\alias{estimate_cwd_load}
\title{Estimate coarse woody debris (1000-hour) fuel loads}
\usage{
estimate_cwd_load(dataset, type, k_value)
}
\arguments{
\item{dataset}{A tidy data frame with a row for each observation, and
columns for the sum of squared diamters for all sound and rotten 1000-hour
fuels along the transect, and the observation-specific fuel coefficients. The
observation-specific coefficients are given by get_1000h_coeffs(), and
depend on the average secant of acute angle and
specific gravity of the various species present in the overstory. Because
the observed data include the sum of squared diameters of intersecting fuels,
we don't need to estimate the quadratic mean diameter of the 1000-hour fuels.}

\item{type}{A string, either 'rotten' or 'sound' decribing which
values to read and write.}

\item{k_value}{Currentlly k_value must equalt 1.234. This is a
unit-conversion constant to relate the units of the observation data
(transect lengths, square centimeters of cross-sectional area per transect)
to the desired results units (metric tons per hectare). Brown's equations
allow for a variety of input and output unit types, with different k values
to translate between them. However, this code currently requires that all
input and output values be in metric - see import_fuels() for more details.}
}
\value{
A numeric vector giving the estimated fuel loading (in Mg / ha) of
the coarse woody debris represented by the observed sum of squared
diameters on the transect.
}
\description{
Coarse woody debris is measured by recording the diameter and decay class
of all 1000-hour fuels (7.63+ cm in diameter).
Van WAgtendonk et al. (1996) and Brown (1974) give equations to estimate
the fuel load (weight / area) represented by these data. For more details,
see the vignette.
}
