% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fuel_loads.R
\name{estimate_fwd_load}
\alias{estimate_fwd_load}
\title{Estimate fine woody debris (1-hour, 10-hour, and 100-hour) fuel loads}
\usage{
estimate_fwd_load(dataset, timelag_class, k_value)
}
\arguments{
\item{dataset}{A tidy data frame with a row for each observation, and
columns for the observed fine woody debris counts and the observation-
specific fuel coefficient. The
observation-specific coefficient is given by get_fwd_coeffs(), and
depends on the average quadratic mean diameter, secant of acute angle, and
specific gravity of the various species present in the overstory.}

\item{timelag_class}{A string, either 'x1h', 'x10h', or 'x100h'. Describes
which columns to read and write.}

\item{k_value}{Currentlly k_value must equalt 1.234. This is a
unit-conversion constant to relate the units of the observation data
(transect lengths, square centimeters of cross-sectional area per transect)
to the desired results units (metric tons per hectare). Brown's equations
allow for a variety of input and output unit types, with different k values
to translate between them. However, this code currently requires that all
input and output values be in metric - see import_fuels() for more details.}
}
\value{
A numeric vector giving the estimated fuel loading (in Mg / ha) of
the fine woody fuel represented by the tallied amount on the transect.
}
\description{
Fine woody debris is measured as tallies by timelag classification along
a transect. The timelag classifications are 1-hour (0 - 0.64cm diameter),
10-hour (0.64 - 2.54 cm diameter), and 100-hour (2.54 - 7.62 cm diameter).
Van WAgtendonk et al. (1996) and Brown (1974) give equations to estimate
the fuel load (weight / area) represented by these tallies. For more details,
see the vignette or the
\href{https://github.com/danfosterfire/Rfuels/blob/master/README.md}{README}.
}
