% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fuel_loads.R
\name{estimate_litterduff_load}
\alias{estimate_litterduff_load}
\title{Estimate litter and duff fuel loads}
\usage{
estimate_litterduff_load(dataset, fuel_type)
}
\arguments{
\item{dataset}{A tidy data frame with a row for each observation, and
columns for the observed litter or duff depth (in cm) and the observation-
specific coefficient between litter/duff depth and fuel load. The
observation-specific coefficient is given by get_litterduff_coeffs()}

\item{fuel_type}{A string, either 'litter', or 'duff'.}
}
\value{
A numeric vector giving the estimated fuel loading (in Mg / ha) of
either litter or duff represented by the observed litter / duff depth on the
transect.
}
\description{
Litter and duff are measured as depths at specific points along a sampling
transect. Van WAgtendonk et al. (1998) developed regressions for litter,
duff, and combined-litter-and-duff loading (kg / m^2) as a function of
depth (cm) for 19 different Sierra Nevada conifer species. See vignette
for details.
}
