% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fuels_coeffs.R
\name{get_1000h_coeffs}
\alias{get_1000h_coeffs}
\title{1000-hour coefficients}
\usage{
get_1000h_coeffs(dataset, species_list, type)
}
\arguments{
\item{dataset}{A tidy data table with a row for each observation (brown's
transect), and columns for the plot identifier ('plot_id'), the inventory
date ('inv_date'), the proportion of basal area occupied by each species
('pBA_species'), the azimuth for the Brown's transect ('azimuth'),
tallies for 1-hour, 10-hour, and 100-hour fuels ('count_1h', 'count_10h',
'count_100h'), transect-average litter and duff depths in centimeters
('duff_depth_cm' and 'litter_depth_cm'), and the sum-of-squared-diameters
for the rotten and sound 1000-hour fuels on the transect ('sum_d2_1000r_cm2'
and 'sum_d2_1000s_cm2').}

\item{species_list}{A characters vector giving the unique species codes
present in the dataset. Generated by calling get_spp_codes(dataset).}

\item{type}{A string, either 'sound' or 'rotten', identifying
the targeted average coefficients.}
}
\value{
The function returns a vector of the product of
the weighted-average SEC and SG for each observation in the dataset.
}
\description{
Van Wagtendonk et al. (1996) tables 3, 6, and give the average quadratic
mean diamter, secant of acute angle, and specific gravity of woody
fuel particles of 21 Sierra Nevada conifers.
Following Stephens et al. (2001), we
estimate the average SEC (secant of acute angle) and SG (specific gravity)
by taking a weighted average of the
species-specific values. The various coefficients are weighted by the
proportion of total plot basal area occupied by each species. Note that
for 1000-hour fuels, we assume that the sum-of-squared-diameters have been
observed and recorded in the data, so we don't need to estimate QMD based
on the species. For more details, see the vignette.
}
