% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fuels_coeffs.R
\name{get_fwd_coeffs}
\alias{get_fwd_coeffs}
\title{1-, 10-, and 100-hour coefficients}
\usage{
get_fwd_coeffs(dataset, species_list, timelag_class)
}
\arguments{
\item{dataset}{A tidy data table with a row for each observation (brown's
transect), and columns for the plot identifier ('plot_id'), the inventory
date ('inv_date'), the proportion of basal area occupied by each species
('pBA_species'), the azimuth for the Brown's transect ('azimuth'),
tallies for 1-hour, 10-hour, and 100-hour fuels ('count_1h', 'count_10h',
'count_100h'), transect-average litter and duff depths in centimeters
('duff_depth_cm' and 'litter_depth_cm'), and the sum-of-squared-diameters
for the rotten and sound 1000-hour fuels on the transect ('sum_d2_1000r_cm2'
and 'sum_d2_1000s_cm2').}

\item{species_list}{A characters vector giving the unique species codes
present in the dataset. Generated by calling get_spp_codes(dataset).}

\item{timelag_class}{A string, ('x1h', 'x10h', or 'x100h') identifying
the type of fine woody debris to give a coefficient for.}
}
\value{
The function returns a vector of the product of
the weighted-average QMD, SEC, and SG for each observation in the dataset.
}
\description{
Van Wagtendonk et al. (1996) tables 3, 6, and give the average quadratic
mean diamter, secant of acute angle, and specific gravity of woody
fuel particles of 21 Sierra Nevada conifers.
Following Stephens et al. (2001), we
estimate the average QMD, SEC, and SG by taking a weighted average of the
species-specific values. The various coefficients  are weighted by the
proportion of total plot basal area occupied by each species. For more
details, see the vignette.
}
