% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fuels_coeffs.R
\name{get_spp_codes}
\alias{get_spp_codes}
\title{Get species codes}
\usage{
get_spp_codes(dataset)
}
\arguments{
\item{dataset}{A tidy dataset with observation ID info and overstory data
(the proportion of basal area occupied by each species,
'pBA_species'). 'dataset' is the result of calling aggregate_treelist() on
the observed treelist data.}
}
\value{
A vector listing the 4-letter codes for each species included in the
dataset, for feeding to the 'get_fuels_coefficients' functions.
}
\description{
Internal function only, see 'estimate_fuel_loads.R' for call. We want a
list of all the species codes included
in the dataset (and including only those species actually
present in the dataset). The tidy dataset returned
by aggregate_treelist() (with a row for each
plot_id:inv_date and a column for the proportion of basal area for each
species) already includes the correct species codes. We want to scan the
column names to see the codes which were present in the dataset.
}
