% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kvals}
\alias{kvals}
\title{Unit conversion constants.}
\format{A data frame with 5 columns and 8 rows:
\describe{
   \item{fuel_diam}{Whether the measurement data describe fuel diameters in
   inches or meters}
   \item{transect_length}{whether the measurement data describe the length of
   the sampling transects in feet or meters}
   \item{volume_fuel_per_area}{the desired output units for describing the
   volume of fuel per unit area (e.g. 'ft^3 / ac", "m^2 / ha", etc.)}
   \item{weight_fuel_per_area}{the desired output units for describing the
   weight of fuel per unit area (e.g. "kg / m^2", "tons / ac", etc.).
   "tons / ac" indicates the North American ton, ie 2,000 pounds. "Tons / ha"
   indicates metric tons, ie 1,000 kilograms.}
   \item{k}{Unitless conversion constant used to map from the input units
   associated with the measurements to the desired output units. For example,
   if fuel diameters are recorded in cm and transect lengths in meters, and
   we want to estimate the fuel load in metric tons per hectare, k = 1.234}

}}
\source{
Values and table format are drawn from
Van Wagner (1982) "Practical Aspects of the Line Intercept Method" (Petawa
National Forestry Institute, Canadian Forestry Service; Chalk River, Ontario
Information Report PI-X-12).
\url{http://www.cfs.nrcan.gc.ca/pubwarehouse/pdfs/6862.pdf}
}
\usage{
kvals
}
\description{
Unit conversion constants used by Brown's Transects equations (variable 'k'
or 'const' depending on the source). These constants are used
to translate between the various units possible for the inputs (fuel
diameters, transect lengths) and outputs (volume of fuel per area,
weight of fuel per area).
}
\keyword{datasets}
