% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litterduff_coeffs}
\alias{litterduff_coeffs}
\title{Litter and Duff Coefficients}
\format{A dataframe with 5 columns and 21 rows:
\describe{
  \item{species}{Common names of the Sierran conifer species included in the
  study}
  \item{spp}{Abbreviated species codes for each species, commonly a 4-letter
  code constructed from the first two names of the genus and the first two
  letters of the specific epithet. These codes are used by Rfuels functions
  and should be used to describe overstory composition in users' input
  data.}
  \item{litter_coeff}{The fitted coefficient for litter weight (kg / m^2)
  as a function of litter depth (cm) for each tree species.}
  \item{duff_coeff}{The fitted coefficient for duff weight (kg / m^2)
  as a function of duff depth (cm) for each tree species.}
  \item{litterduff_coeff}{The fitted coefficient for litter and duff
  combined weight (kg / m^2) as a function of the combined depth of the
  litter and duff layers (cm) for each tree species. Note that fuelbed
  weights predicted by the "litter and duff" coefficient are NOT equal to
  the summed predictions for litter and duff individually.}
}}
\source{
Table 7, Van Wagtendonk, Benedict, and Sydoriak (1998) West. J. Appl.
For. 13(3):73-84 \url{https://academic.oup.com/wjaf/article/13/3/73/4663215}
}
\usage{
litterduff_coeffs
}
\description{
Regression coefficients for litter, Duff, and litter&Duff weight (kg/m^2) as
a function of fuelbed depth (cm) for the 19 Sierra Nevada conifers listed.
Fuels were collected from 4 pure stands of each of the 22 species listed to
fit the regressions. Each regression went through the origin, was based on 80
observations, and was significant at the 0.05 level.
}
\keyword{datasets}
