% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fuels_coeffs.R
\name{lookup_vw}
\alias{lookup_vw}
\title{Lookup Van Wagtendonk Values}
\usage{
lookup_vw(constants, species, value)
}
\arguments{
\item{constants}{A dataframe with the relevant constants table. See
data-raw/build_constants_tables.R for construction of the dataframes.}

\item{species}{String with the desired species code, drawn from the set of
codes included in the 'spp' columns included in the 'constants' dataframe.}

\item{value}{String giving the desired value (column) from the table.}
}
\description{
Internal only. Helper function to fetch a specific constant from the
tables reported by Van Wagtendonk et al. (1996) and (1998) based on the
type of constant (which table to check, e.g. QMD_cm or SEC), the species
(which row to check), and the specific value
(which column, e.g. 1-hour or 1-hour ).
}
