% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QMDcm}
\alias{QMDcm}
\title{Average squared quadratic mean diameter (QMD^2) by fuel size class for 19
Sierra Nevada conifers}
\format{A dataframe with 6 columns and 21 rows:
\describe{
  \item{Species}{The latin name of each of the 19 conifer species included
  in the study.}
  \item{spp}{Abbreviated species codes for each species, commonly a 4-letter
  code constructed from the first two names of the genus and the first two
  letters of the specific epithet. These codes are used by Rfuels functions
  and should be used to describe overstory composition in users' input
  data.}
  \item{x1h}{The average squared quadratic mean diameter (in cm^2) of a
  1-hour fuel (a woody particle between 0cm and 0.64cm in diameter at the
  point of intersection with the sampling transect) for each species.}
  \item{x10h}{The average squared quadratic mean diameter (in cm^2) of a
  10-hour fuel (a woody particle between 0.64cm and 2.54cm in diameter
  at the point of intersection with the sampling transect) for each species.}
  \item{x100h}{The average squared quadratic mean diameter (in cm^2) of a
  100-hour fuel (a woody particle between 2.54cm and 7.62cm in diameter at
  the point of intersection with the sampling transect) for each species.}
  \item{x1000h}{The average squared quadratic mean diameter (in cm^2) of a
  1000-hour fuel (a woody particle greater than 7.62cm in diameter at the
  point of intersection with the sampling transect) for each species. Most
  sampling protocols call for direct measurement of the diameter for any
  1000-hour fuels intersecting the transect; where the actual diameter is
  known it is not necessary to rely on these average values.}
}}
\source{
van Wagtendonk, Benedict, and Sydoriak (1996) "Physical Properties of
Woody Fuel Particles of Sierra Nevada Conifers" (Int. J. Wildland Fire 6(3):
117-123); Table 3. \url{https://doi.org/10.1071/WF9960117}
}
\usage{
QMDcm
}
\description{
When using Brown's Transects, fine woody fuels are recorded as tallies
of the number of fuel particales (sticks) crossing a sampling transect. Fine
woody fuels are stratified into size classes of 0-0.64cm (1-hour fuels),
0.64-2.54cm (10-hour fuels), 2.54-7.62cm (100-hour fuels), and 7.62+ cm
(1000-hour fuels). (These "timelag classifications" are related to the amount
of time for fuels of each size to gain or lose moisture in response to changes
in the relative humidity of the surrounding air - smaller fuels will respond
more quickly to changes in the moisture of the surrounding environment.)
Van Wagtendonk et al. (1996) sampled fuels from pure stands of each tree
species, and recorded summary statistics for each species and timelag
class. From the paper: "Diameter was calculated for each fuel particle by
averaging two perpendicular measurements made at the point the particle
intersected with the sampling plane. Squared quadratic mean diameter
was calculated by squaring each particle's average diameter and taking the
average of the sum for each stand and size class."
}
\keyword{datasets}
