% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEC}
\alias{SEC}
\title{Average secand of acute angles of inclination of nonhorizontal particles by
fuel size class for 19 Sierra Nevada conifers}
\format{A dataframe with 6 columns and 21 rows:
\describe{
  \item{Species}{The latin name of each of the 19 conifer species included
  in the study.}
  \item{spp}{Abbreviated species codes for each species, commonly a 4-letter
  code constructed from the first two names of the genus and the first two
  letters of the specific epithet. These codes are used by Rfuels functions
  and should be used to describe overstory composition in users' input
  data.}
  \item{x1h}{The average secant of actute angle of inclination of a
  1-hour fuel (a woody particle between 0cm and 0.64cm in diameter at the
  point of intersection with the sampling transect) for each species.}
  \item{x10h}{The average secant of actute angle of inclination of a
  10-hour fuel (a woody particle between 0.64cm and 2.54cm in diameter
  at the point of intersection with the sampling transect) for each species.}
  \item{x100h}{The average secant of actute angle of inclination of a
  100-hour fuel (a woody particle between 2.54cm and 7.62cm in diameter at
  the point of intersection with the sampling transect) for each species.}
  \item{x1000h}{The average secant of actute angle of inclination of a
  1000-hour fuel (a woody particle greater than 7.62cm in diameter at the
  point of intersection with the sampling transect) for each species.}
}}
\source{
van Wagtendonk, Benedict, and Sydoriak (1996) "Physical Properties of
Woody Fuel Particles of Sierra Nevada Conifers" (Int. J. Wildland Fire 6(3):
117-123); Table 6. \url{https://doi.org/10.1071/WF9960117}
}
\usage{
SEC
}
\description{
Brown's equations for estimating fuel loads from sampling transects call for
the "secant of acute angle," which accounts for the fact that most fuel
particles intersecting the transect do not lie exacatly horizontally when
estimating the volume (or weight) of fuels intersected by the transect. Van
Wagtendonk et al. report that "the acute angle that the intersected particle
formed with a horizontal plane was measured using a protractor level to the
nearest five degrees" and give average values for the secant of this angle
for each species and timelag class.
}
\keyword{datasets}
