% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SG}
\alias{SG}
\title{Average secand of acute angles of inclination of nonhorizontal particles by
fuel size class for 19 Sierra Nevada conifers}
\format{A dataframe with 6 columns and 21 rows:
\describe{
  \item{Species}{The latin name of each of the 19 conifer species included
  in the study.}
  \item{spp}{Abbreviated species codes for each species, commonly a 4-letter
  code constructed from the first two names of the genus and the first two
  letters of the specific epithet. These codes are used by Rfuels functions
  and should be used to describe overstory composition in users' input
  data.}
  \item{x1h}{The average specific gravity of a
  1-hour fuel (a woody particle between 0cm and 0.64cm in diameter at the
  point of intersection with the sampling transect) for each species.}
  \item{x10h}{The average specific gravity of a
  10-hour fuel (a woody particle between 0.64cm and 2.54cm in diameter
  at the point of intersection with the sampling transect) for each species.}
  \item{x100h}{The average specific gravity of a
  100-hour fuel (a woody particle between 2.54cm and 7.62cm in diameter at
  the point of intersection with the sampling transect) for each species.}
  \item{x1000h}{The average specific gravity of a
  1000-hour fuel (a woody particle greater than 7.62cm in diameter at the
  point of intersection with the sampling transect) for each species.}
}}
\source{
van Wagtendonk, Benedict, and Sydoriak (1996) "Physical Properties of
Woody Fuel Particles of Sierra Nevada Conifers" (Int. J. Wildland Fire 6(3):
117-123); Table 8. \url{https://doi.org/10.1071/WF9960117}
}
\usage{
SG
}
\description{
Brown's equations for estimating fuel loads from sampling transects call for
the specific gravity of the fuel particles. Van Wagtendonk et al. used a
Kraus Jolly specific gravity balance (comparing the weight of a sampling
in air to its weight immersed in water, sensu Eberbach 1979) to
determine the specific gravity of a subset of their woody fuel samples
for each species, and timelag class (with sound and rotten particles analyzed
separately).
}
\keyword{datasets}
