% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fuel_loads.R
\name{estimate_fuel_loads}
\alias{estimate_fuel_loads}
\title{Estimate fuel loads}
\usage{
estimate_fuel_loads(fuels_data, trees_data,
  results_type = "results_only")
}
\arguments{
\item{fuels_data}{The dataframe with the Brown's transect observations
describing the fuels. See the vignette and/or README for more info.}

\item{trees_data}{The dataframe with the observations of individual trees
(aka the treelist). See the vignette and/or README for more info.}

\item{results_type}{Either 'full', 'results_only', or 'fuels_only'. Sets the
verbosity of the returned results dataframe. 'full' includes observation
ID information, all directly observed values, intermediate calculations,
overstory species composition, and the fuel load estimates.
'results_only' returns the observation ID
information, the overstory species composition, and the fuel load estimates.
'fuels_only' returns the observation ID information and the fuel load
estimates.}
}
\value{
A tidy data frame, with a row for each observation (where an
observation is a Brown's transect on a specific date), and columns for
the plot location (plot_id), the inventory date (inv_date), the transect
azimuth (azimuth), and the fuel load estimated for various subcategories
of surface and ground fuels (litter, duff, 1-hour, 10-hour, 100-hour, and
1000-hour fuels). Users may opt to include columns for the directly observed
measurements, intermediate values, and the overstory species composition
using the parameter 'results_type'.
}
\description{
A top-level wrapper function to estimate fuel loads from Brown's
transect data and a treelist describing the local
overstory for each observation. For details, see the vignette.
}
